pyproject.toml
[tool.poetry]
name = "django-basic-template"
version = "0.1.0"
description = ""
authors = ["hwangtate <ghkdxodud10253504@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
black = "^24.8.0"
python-dotenv = "^1.0.1"
mysqlclient = "^2.2.4"
djangorestframework = "^3.15.2"
django-session-timeout = "^0.1.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
